+++ includes/class-partition-node.php
@@ -414,9 +414,9 @@ class Partition_Node extends Timer_Node {
 			// Heartbeat cadence = stale_timeout/3 ms; three ticks per stale window.
 			$this->heartbeat_timer = new Timer_Node();
 			$this->heartbeat_timer->name( "{$this->name}:heartbeat" );
+			$this->heartbeat_timer->arguments( (string) \intdiv( $stale_timeout * 1000, 3 ) );
 			$this->heartbeat_timer->sink( $this->write_lock );
-			$this->heartbeat_timer->set_key( 'heartbeat' );
-			$this->heartbeat_timer->set_timer( $stale_timeout * 1000 / 3 );
+			$this->heartbeat_timer->key( 'heartbeat' );
 			$this->heartbeat_timer->patron( $this );
 		}